requirements.txt
openai
flask
pinecone
pinecone-client
PyGithub
tiktoken
nltk
markdown
gitpython